tx/cos_classify.rs: colocate 24 tests with production (#984 P3 Phase 2a)
Moves all 24 tests targeting cos_classify functions
(clone_prepared_request_for_cos, cos_queue_accepts_prepared,
demote_prepared_cos_queue_to_local, prepare_local_request_for_cos,
resolve_cached_cos_tx_selection, resolve_cos_queue_id/idx,
resolve_cos_tx_selection) from tx/mod.rs::tests to a new mod tests
in tx/cos_classify.rs.

The new tests block imports super::*, test_support::*, plus 11
crate-level Snapshot types (ClassOfServiceSnapshot, the four CoS*
classifier snapshots, scheduler-map/scheduler/forwarding-class
snapshots, FirewallFilter + Term snapshots).

Also drops:
- The 11-name use crate::{...Snapshot} block from
  tx/mod.rs::tests (no longer needed there).
- 3 stale cfg-test cos_classify pin imports
  (clone_prepared_request_for_cos, prepare_local_request_for_cos,
  resolve_cos_queue_idx). Only cos_queue_accepts_prepared and
  demote_prepared_cos_queue_to_local stay (still used by
  remaining tx/mod.rs::tests).
- Refresh banner comment that said '5 helpers' to reflect the
  remaining 2.

Rebased onto current master after #1001 (cross_binding) and
#1003 (ecn) merged. cargo test --bins 865/0/2.
