+++ userspace-dp/src/afxdp/tx/mod.rs
@@ -73,13 +73,13 @@ pub(in crate::afxdp) use cos_classify::cos_queue_dscp_rewrite;
 // Private import (NOT re-export — E0364 if pub(super) re-export of
 // pub(super) source): drain.rs reaches it via `use super::*;`.
 use cos_classify::enqueue_prepared_into_cos;
-// cfg-test imports for 5 helpers directly pinned in tx/mod.rs::tests.
+// cfg-test imports for 2 helpers still pinned in tx/mod.rs::tests
+// (the other 3 cos_classify helper tests moved to
+// cos_classify.rs::tests in #984 P3 phase 2a).
 #[cfg(test)]
 use cos_classify::{
-    clone_prepared_request_for_cos, cos_queue_accepts_prepared,
-    demote_prepared_cos_queue_to_local, prepare_local_request_for_cos,
-    resolve_cos_queue_idx,
-};
+    cos_queue_accepts_prepared,
+    demote_prepared_cos_queue_to_local, };
 
 
 
@@ -114,11 +114,12 @@ use cos_classify::{
 // Production code uses the entry points re-exported from
 // cos/mod.rs (marker fns + flow-hash + admission gates +
 // flow-fair promotion entry + token-bucket helpers + queue-ops
-// fns). The codepoint masks + ECN parser + per-family ECN helpers
-// are referenced only by `tx::tests` (admission tests + ECN unit
-// tests that stay here for Phase 1). Their imports are gated
-// behind `#[cfg(test)]` to avoid `unused_imports` warnings in
-// non-test builds (Copilot review on PR #976).
+// fns). The remaining ECN constants/thresholds are referenced
+// only by admission tests in `tx::tests`; the ECN parser and
+// per-family ECN helpers (`ethernet_l3`, `mark_ecn_ce_*`) moved
+// with their unit tests to `cos/ecn.rs` in #984 P3 phase 2c.
+// Imports are gated behind `#[cfg(test)]` to avoid
+// `unused_imports` warnings in non-test builds.
 use super::cos::{
     apply_cos_admission_ecn_policy, cos_flow_aware_buffer_limit, cos_flow_bucket_index,
     cos_item_flow_key, cos_queue_drain_all, cos_queue_flow_share_limit,
@@ -128,8 +129,6 @@ use super::cos::{
     redirect_prepared_cos_request_to_owner_binding,
     resolve_local_routing_decision, LocalRoutingDecision, Step1Action,
 };
-#[cfg(test)]
-use super::cos::ecn::{ethernet_l3, mark_ecn_ce_ipv4, mark_ecn_ce_ipv6, EthernetL3};
 #[cfg(test)]
 use super::cos::{
     account_cos_queue_flow_dequeue, account_cos_queue_flow_enqueue,
@@ -142,12 +141,12 @@ use super::cos::{
     drain_exact_local_fifo_items_to_scratch, drain_exact_local_items_to_scratch_flow_fair,
     drain_exact_prepared_fifo_items_to_scratch,
     drain_exact_prepared_items_to_scratch_flow_fair, estimate_cos_queue_wakeup_tick,
-    maybe_mark_ecn_ce, release_exact_local_scratch_frames, release_exact_prepared_scratch,
+    release_exact_local_scratch_frames, release_exact_prepared_scratch,
     select_cos_guarantee_batch, select_cos_guarantee_batch_with_fast_path,
     select_cos_surplus_batch, select_exact_cos_guarantee_queue_with_fast_path,
     select_nonexact_cos_guarantee_batch, settle_exact_local_fifo_submission,
     settle_exact_local_scratch_submission_flow_fair, settle_exact_prepared_fifo_submission, COS_ECN_MARK_THRESHOLD_DEN, COS_ECN_MARK_THRESHOLD_NUM,
-    COS_FLOW_FAIR_MIN_SHARE_BYTES, ECN_CE, ECN_ECT_0, ECN_ECT_1, ECN_MASK, ECN_NOT_ECT,
+    COS_FLOW_FAIR_MIN_SHARE_BYTES, ECN_CE, ECN_ECT_0, ECN_MASK, ECN_NOT_ECT,
     V_MIN_CONSECUTIVE_SKIP_HARD_CAP, V_MIN_SUSPENSION_BATCHES,
 };
 // #956 P1: TX-completion + timer-wheel items reached by
@@ -191,12 +190,6 @@ use super::cos::{
 mod tests {
     use super::*;
     use super::test_support::*;
-    use crate::{
-        ClassOfServiceSnapshot, CoSDSCPClassifierEntrySnapshot, CoSDSCPClassifierSnapshot,
-        CoSForwardingClassSnapshot, CoSIEEE8021ClassifierEntrySnapshot,
-        CoSIEEE8021ClassifierSnapshot, CoSSchedulerMapEntrySnapshot, CoSSchedulerMapSnapshot,
-        CoSSchedulerSnapshot, FirewallFilterSnapshot, FirewallTermSnapshot,
-    };
 
 
 
@@ -221,25 +214,6 @@ mod tests {
 
 
 
-    #[test]
-    fn resolve_cos_queue_idx_rejects_explicit_queue_miss() {
-        let root = test_cos_runtime_with_queues(
-            10_000_000,
-            vec![CoSQueueConfig {
-                queue_id: 5,
-                forwarding_class: "best-effort".into(),
-                priority: 5,
-                transmit_rate_bytes: 10_000_000,
-                exact: false,
-                surplus_weight: 1,
-                buffer_bytes: COS_MIN_BURST_BYTES,
-                dscp_rewrite: None,
-            }],
-        );
-
-        assert_eq!(resolve_cos_queue_idx(&root, Some(4)), None);
-        assert_eq!(resolve_cos_queue_idx(&root, None), Some(0));
-    }
 
 
     #[test]
@@ -513,162 +487,10 @@ mod tests {
 
 
 
-    #[test]
-    fn clone_prepared_request_for_cos_returns_local_copy_with_metadata() {
-        let mut area = MmapArea::new(4096).expect("mmap");
-        let payload = [0xde, 0xad, 0xbe, 0xef];
-        area.slice_mut(128, payload.len())
-            .expect("slice")
-            .copy_from_slice(&payload);
-        let req = PreparedTxRequest {
-            offset: 128,
-            len: payload.len() as u32,
-            recycle: PreparedTxRecycle::FreeTxFrame,
-            expected_ports: Some((1111, 2222)),
-            expected_addr_family: libc::AF_INET6 as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: Some(SessionKey {
-                addr_family: libc::AF_INET6 as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V6(Ipv6Addr::LOCALHOST),
-                dst_ip: IpAddr::V6(Ipv6Addr::LOCALHOST),
-                src_port: 1111,
-                dst_port: 2222,
-            }),
-            egress_ifindex: 80,
-            cos_queue_id: Some(4),
-            dscp_rewrite: Some(46),
-        };
 
-        let local = clone_prepared_request_for_cos(&area, &req).expect("local copy");
 
-        assert_eq!(local.bytes, payload);
-        assert_eq!(local.expected_ports, Some((1111, 2222)));
-        assert_eq!(local.expected_addr_family, libc::AF_INET6 as u8);
-        assert_eq!(local.expected_protocol, PROTO_TCP);
-        assert_eq!(local.egress_ifindex, 80);
-        assert_eq!(local.cos_queue_id, Some(4));
-        assert_eq!(local.dscp_rewrite, Some(46));
-        assert_eq!(
-            local
-                .flow_key
-                .as_ref()
-                .map(|key| (key.src_port, key.dst_port)),
-            Some((1111, 2222))
-        );
-    }
 
-    #[test]
-    fn clone_prepared_request_for_cos_rejects_out_of_range_offset() {
-        let area = MmapArea::new(256).expect("mmap");
-        let req = PreparedTxRequest {
-            offset: 1024,
-            len: 64,
-            recycle: PreparedTxRecycle::FreeTxFrame,
-            expected_ports: None,
-            expected_addr_family: libc::AF_INET as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: None,
-            egress_ifindex: 80,
-            cos_queue_id: Some(4),
-            dscp_rewrite: None,
-        };
 
-        assert!(clone_prepared_request_for_cos(&area, &req).is_none());
-    }
-
-    #[test]
-    fn prepare_local_request_for_cos_materializes_prepared_frame() {
-        let area = MmapArea::new(4096).expect("mmap");
-        let mut free_tx_frames = VecDeque::from([128]);
-        let req = TxRequest {
-            bytes: vec![0xde, 0xad, 0xbe, 0xef],
-            expected_ports: Some((1111, 2222)),
-            expected_addr_family: libc::AF_INET6 as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: Some(SessionKey {
-                addr_family: libc::AF_INET6 as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V6(Ipv6Addr::LOCALHOST),
-                dst_ip: IpAddr::V6(Ipv6Addr::LOCALHOST),
-                src_port: 1111,
-                dst_port: 2222,
-            }),
-            egress_ifindex: 80,
-            cos_queue_id: Some(5),
-            dscp_rewrite: Some(46),
-        };
-
-        let prepared =
-            prepare_local_request_for_cos(&area, &mut free_tx_frames, req).expect("prepared");
-
-        assert_eq!(prepared.offset, 128);
-        assert_eq!(prepared.len, 4);
-        assert_eq!(prepared.recycle, PreparedTxRecycle::FreeTxFrame);
-        assert_eq!(prepared.expected_ports, Some((1111, 2222)));
-        assert_eq!(prepared.egress_ifindex, 80);
-        assert_eq!(prepared.cos_queue_id, Some(5));
-        assert_eq!(prepared.dscp_rewrite, Some(46));
-        assert!(free_tx_frames.is_empty());
-        assert_eq!(area.slice(128, 4).expect("slice"), [0xde, 0xad, 0xbe, 0xef]);
-    }
-
-    #[test]
-    fn prepare_local_request_for_cos_falls_back_when_no_free_tx_frame_exists() {
-        let area = MmapArea::new(4096).expect("mmap");
-        let mut free_tx_frames = VecDeque::new();
-        let req = TxRequest {
-            bytes: vec![1, 2, 3, 4],
-            expected_ports: None,
-            expected_addr_family: libc::AF_INET as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: None,
-            egress_ifindex: 80,
-            cos_queue_id: Some(5),
-            dscp_rewrite: None,
-        };
-
-        let req = match prepare_local_request_for_cos(&area, &mut free_tx_frames, req) {
-            Ok(_) => panic!("must fall back to local"),
-            Err(req) => req,
-        };
-
-        assert_eq!(req.bytes, [1, 2, 3, 4]);
-        assert!(free_tx_frames.is_empty());
-    }
-
-    #[test]
-    fn cos_queue_accepts_prepared_when_queue_is_prepared_only() {
-        let mut root = test_cos_runtime_with_queues(
-            10_000_000_000 / 8,
-            vec![CoSQueueConfig {
-                queue_id: 5,
-                forwarding_class: "iperf-b".into(),
-                priority: 5,
-                transmit_rate_bytes: 10_000_000_000 / 8,
-                exact: true,
-                surplus_weight: 1,
-                buffer_bytes: COS_MIN_BURST_BYTES,
-                dscp_rewrite: None,
-            }],
-        );
-        root.queues[0]
-            .items
-            .push_back(CoSPendingTxItem::Prepared(PreparedTxRequest {
-                offset: 64,
-                len: 1500,
-                recycle: PreparedTxRecycle::FreeTxFrame,
-                expected_ports: None,
-                expected_addr_family: libc::AF_INET6 as u8,
-                expected_protocol: PROTO_TCP,
-                flow_key: None,
-                egress_ifindex: 80,
-                cos_queue_id: Some(5),
-                dscp_rewrite: None,
-            }));
-
-        assert!(cos_queue_accepts_prepared(&root, Some(5)));
-    }
 
     #[test]
     fn cos_queue_rejects_prepared_once_local_items_enter_queue() {
@@ -720,291 +542,8 @@ mod tests {
         assert!(!cos_queue_accepts_prepared(&root, Some(5)));
     }
 
-    #[test]
-    fn demote_prepared_cos_queue_to_local_recycles_frames_and_blocks_prepared_appends() {
-        let area = MmapArea::new(4096).expect("mmap");
-        unsafe { area.slice_mut_unchecked(64, 4) }
-            .expect("frame")
-            .copy_from_slice(&[0xde, 0xad, 0xbe, 0xef]);
-        unsafe { area.slice_mut_unchecked(128, 4) }
-            .expect("frame")
-            .copy_from_slice(&[0xca, 0xfe, 0xba, 0xbe]);
 
-        let mut root = test_cos_runtime_with_queues(
-            10_000_000_000 / 8,
-            vec![CoSQueueConfig {
-                queue_id: 5,
-                forwarding_class: "iperf-b".into(),
-                priority: 5,
-                transmit_rate_bytes: 10_000_000_000 / 8,
-                exact: true,
-                surplus_weight: 1,
-                buffer_bytes: COS_MIN_BURST_BYTES,
-                dscp_rewrite: None,
-            }],
-        );
-        root.queues[0]
-            .items
-            .push_back(CoSPendingTxItem::Prepared(PreparedTxRequest {
-                offset: 64,
-                len: 4,
-                recycle: PreparedTxRecycle::FreeTxFrame,
-                expected_ports: Some((1111, 5202)),
-                expected_addr_family: libc::AF_INET6 as u8,
-                expected_protocol: PROTO_TCP,
-                flow_key: None,
-                egress_ifindex: 80,
-                cos_queue_id: Some(5),
-                dscp_rewrite: None,
-            }));
-        root.queues[0]
-            .items
-            .push_back(CoSPendingTxItem::Prepared(PreparedTxRequest {
-                offset: 128,
-                len: 4,
-                recycle: PreparedTxRecycle::FillOnSlot(7),
-                expected_ports: Some((1112, 5202)),
-                expected_addr_family: libc::AF_INET6 as u8,
-                expected_protocol: PROTO_TCP,
-                flow_key: None,
-                egress_ifindex: 80,
-                cos_queue_id: Some(5),
-                dscp_rewrite: None,
-            }));
 
-        let mut free_tx_frames = VecDeque::from([512]);
-        let mut pending_fill_frames = VecDeque::new();
-        assert!(demote_prepared_cos_queue_to_local(
-            &area,
-            &mut free_tx_frames,
-            &mut pending_fill_frames,
-            7,
-            &mut root,
-            Some(5),
-        ));
-
-        let items = root.queues[0]
-            .items
-            .iter()
-            .map(|item| match item {
-                CoSPendingTxItem::Local(req) => req.bytes.clone(),
-                CoSPendingTxItem::Prepared(_) => panic!("prepared item should be demoted"),
-            })
-            .collect::<Vec<_>>();
-        assert_eq!(
-            items,
-            vec![vec![0xde, 0xad, 0xbe, 0xef], vec![0xca, 0xfe, 0xba, 0xbe]]
-        );
-        assert_eq!(free_tx_frames, VecDeque::from([512, 64]));
-        assert_eq!(pending_fill_frames, VecDeque::from([128]));
-        assert!(!cos_queue_accepts_prepared(&root, Some(5)));
-    }
-
-    /// #926: regression test for the success-path
-    /// queue_vtime / head-finish preservation. Prepared items
-    /// across multiple flows are queued, demoted to Local, and
-    /// the MQFQ frontier (queue_vtime + per-bucket head/tail
-    /// finish-times) MUST be unchanged. A new flow Y enqueued
-    /// immediately after demotion MUST anchor at a finish-time
-    /// that respects the demoted backlog's frontier — i.e. Y
-    /// cannot jump ahead of the demoted backlog.
-    #[test]
-    fn demote_prepared_cos_queue_to_local_preserves_mqfq_frontier() {
-        let area = MmapArea::new(4096).expect("mmap");
-        unsafe { area.slice_mut_unchecked(64, 4) }
-            .expect("frame")
-            .copy_from_slice(&[0xde, 0xad, 0xbe, 0xef]);
-        unsafe { area.slice_mut_unchecked(128, 4) }
-            .expect("frame")
-            .copy_from_slice(&[0xca, 0xfe, 0xba, 0xbe]);
-
-        let mut root = test_cos_runtime_with_queues(
-            10_000_000_000 / 8,
-            vec![CoSQueueConfig {
-                queue_id: 4,
-                forwarding_class: "iperf-a".into(),
-                priority: 5,
-                transmit_rate_bytes: 1_000_000_000 / 8,
-                exact: true,
-                surplus_weight: 1,
-                buffer_bytes: 128 * 1024,
-                dscp_rewrite: None,
-            }],
-        );
-        let queue = &mut root.queues[0];
-        queue.flow_fair = true;
-        queue.flow_hash_seed = 0;
-
-        // Two distinct flows, each one Prepared item. Bucket
-        // indices computed under flow_hash_seed=0 for use in
-        // post-demote frontier assertions.
-        let key_a = test_session_key(8001, 5201);
-        let key_b = test_session_key(8002, 5201);
-        let bucket_a = cos_flow_bucket_index(0, Some(&key_a));
-        let bucket_b = cos_flow_bucket_index(0, Some(&key_b));
-        assert_ne!(
-            bucket_a, bucket_b,
-            "test setup: ports 8001/8002 must hash to distinct buckets"
-        );
-
-        cos_queue_push_back(
-            queue,
-            CoSPendingTxItem::Prepared(PreparedTxRequest {
-                offset: 64,
-                len: 1500,
-                recycle: PreparedTxRecycle::FreeTxFrame,
-                expected_ports: None,
-                expected_addr_family: libc::AF_INET as u8,
-                expected_protocol: PROTO_TCP,
-                flow_key: Some(key_a.clone()),
-                egress_ifindex: 42,
-                cos_queue_id: Some(4),
-                dscp_rewrite: None,
-            }),
-        );
-        cos_queue_push_back(
-            queue,
-            CoSPendingTxItem::Prepared(PreparedTxRequest {
-                offset: 128,
-                len: 1500,
-                recycle: PreparedTxRecycle::FreeTxFrame,
-                expected_ports: None,
-                expected_addr_family: libc::AF_INET as u8,
-                expected_protocol: PROTO_TCP,
-                flow_key: Some(key_b.clone()),
-                egress_ifindex: 42,
-                cos_queue_id: Some(4),
-                dscp_rewrite: None,
-            }),
-        );
-
-        // Snapshot pre-demote MQFQ frontier.
-        let pre_vtime = queue.queue_vtime;
-        let pre_head_a = queue.flow_bucket_head_finish_bytes[bucket_a];
-        let pre_head_b = queue.flow_bucket_head_finish_bytes[bucket_b];
-        let pre_tail_a = queue.flow_bucket_tail_finish_bytes[bucket_a];
-        let pre_tail_b = queue.flow_bucket_tail_finish_bytes[bucket_b];
-        assert!(pre_head_a > 0);
-        assert!(pre_head_b > 0);
-
-        // Demote (success path).
-        let mut free_tx_frames = VecDeque::from([512]);
-        let mut pending_fill_frames = VecDeque::new();
-        assert!(demote_prepared_cos_queue_to_local(
-            &area,
-            &mut free_tx_frames,
-            &mut pending_fill_frames,
-            7,
-            &mut root,
-            Some(4),
-        ));
-
-        let queue = &mut root.queues[0];
-
-        // Frontier MUST be unchanged across the success path.
-        assert_eq!(
-            queue.queue_vtime, pre_vtime,
-            "#926 regression: queue_vtime must be preserved across \
-             demote success path. Pre={pre_vtime} post={}",
-            queue.queue_vtime
-        );
-        assert_eq!(
-            queue.flow_bucket_head_finish_bytes[bucket_a], pre_head_a,
-            "#926: head_finish[A] must be preserved (pre={pre_head_a})"
-        );
-        assert_eq!(
-            queue.flow_bucket_head_finish_bytes[bucket_b], pre_head_b,
-            "#926: head_finish[B] must be preserved (pre={pre_head_b})"
-        );
-        assert_eq!(
-            queue.flow_bucket_tail_finish_bytes[bucket_a], pre_tail_a,
-            "#926: tail_finish[A] must be preserved"
-        );
-        assert_eq!(
-            queue.flow_bucket_tail_finish_bytes[bucket_b], pre_tail_b,
-            "#926: tail_finish[B] must be preserved"
-        );
-
-        // Items now Local. flow_fair=true stores items in
-        // per-bucket VecDeques at `flow_bucket_items[bucket]`,
-        // not in `queue.items`.
-        let mut total_items = 0;
-        for bucket in [bucket_a, bucket_b] {
-            for item in queue.flow_bucket_items[bucket].iter() {
-                assert!(
-                    matches!(item, CoSPendingTxItem::Local(_)),
-                    "demote should convert Prepared → Local"
-                );
-                total_items += 1;
-            }
-        }
-        assert_eq!(total_items, 2);
-
-        // The frontier-preservation assertions above are the
-        // load-bearing test (Codex code review caught that an
-        // earlier "Y does not jump ahead" assertion was
-        // logically muddled — without the fix, the four
-        // assert_eq calls already FAIL at the queue_vtime / head /
-        // tail checks; demote_prepared without snapshot/restore
-        // leaves queue_vtime=3000 and head_a=head_b=4500, all
-        // mismatching the captured pre-state). The Y-anchor
-        // behavior at this scenario is identical with-or-without
-        // the fix (Y is small enough to anchor below A/B in
-        // both cases) so it's not a useful gate.
-    }
-
-    #[test]
-    fn demote_prepared_cos_queue_to_local_skips_non_exact_queue() {
-        let area = MmapArea::new(4096).expect("mmap");
-        unsafe { area.slice_mut_unchecked(64, 4) }
-            .expect("frame")
-            .copy_from_slice(&[1, 2, 3, 4]);
-
-        let mut root = test_cos_runtime_with_queues(
-            10_000_000_000 / 8,
-            vec![CoSQueueConfig {
-                queue_id: 5,
-                forwarding_class: "iperf-b".into(),
-                priority: 5,
-                transmit_rate_bytes: 10_000_000_000 / 8,
-                exact: false,
-                surplus_weight: 1,
-                buffer_bytes: COS_MIN_BURST_BYTES,
-                dscp_rewrite: None,
-            }],
-        );
-        root.queues[0]
-            .items
-            .push_back(CoSPendingTxItem::Prepared(PreparedTxRequest {
-                offset: 64,
-                len: 4,
-                recycle: PreparedTxRecycle::FreeTxFrame,
-                expected_ports: None,
-                expected_addr_family: libc::AF_INET6 as u8,
-                expected_protocol: PROTO_TCP,
-                flow_key: None,
-                egress_ifindex: 80,
-                cos_queue_id: Some(5),
-                dscp_rewrite: None,
-            }));
-
-        let mut free_tx_frames = VecDeque::new();
-        let mut pending_fill_frames = VecDeque::new();
-        assert!(!demote_prepared_cos_queue_to_local(
-            &area,
-            &mut free_tx_frames,
-            &mut pending_fill_frames,
-            7,
-            &mut root,
-            Some(5),
-        ));
-        assert!(matches!(
-            root.queues[0].items.front(),
-            Some(CoSPendingTxItem::Prepared(_))
-        ));
-        assert!(free_tx_frames.is_empty());
-        assert!(pending_fill_frames.is_empty());
-    }
 
     #[test]
     fn drain_exact_local_fifo_items_to_scratch_keeps_queue_until_commit() {
@@ -1692,1431 +1231,20 @@ mod tests {
         ));
     }
 
-    #[test]
-    fn resolve_cos_queue_id_prefers_egress_output_filter_forwarding_class() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![
-                InterfaceSnapshot {
-                    name: "reth1.0".into(),
-                    ifindex: 101,
-                    parent_ifindex: 5,
-                    vlan_id: 0,
-                    hardware_addr: "02:bf:72:00:61:01".into(),
-                    filter_input_v4: "cos-classify".into(),
-                    ..Default::default()
-                },
-                InterfaceSnapshot {
-                    name: "reth0.0".into(),
-                    ifindex: 202,
-                    hardware_addr: "02:bf:72:00:80:08".into(),
-                    filter_output_v4: "wan-classify".into(),
-                    cos_shaping_rate_bytes_per_sec: 10_000_000,
-                    cos_shaping_burst_bytes: 256_000,
-                    cos_scheduler_map: "wan-map".into(),
-                    ..Default::default()
-                },
-            ],
-            filters: vec![
-                FirewallFilterSnapshot {
-                    name: "cos-classify".into(),
-                    family: "inet".into(),
-                    terms: vec![FirewallTermSnapshot {
-                        name: "voice".into(),
-                        protocols: vec!["tcp".into()],
-                        destination_ports: vec!["443".into()],
-                        action: "accept".into(),
-                        forwarding_class: "best-effort".into(),
-                        ..Default::default()
-                    }],
-                },
-                FirewallFilterSnapshot {
-                    name: "wan-classify".into(),
-                    family: "inet".into(),
-                    terms: vec![FirewallTermSnapshot {
-                        name: "voice".into(),
-                        protocols: vec!["tcp".into()],
-                        destination_ports: vec!["443".into()],
-                        action: "accept".into(),
-                        forwarding_class: "expedited-forwarding".into(),
-                        ..Default::default()
-                    }],
-                },
-            ],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 0,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "expedited-forwarding".into(),
-                        queue: 1,
-                    },
-                ],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 4_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "ef-sched".into(),
-                        transmit_rate_bytes: 6_000_000,
-                        transmit_rate_exact: false,
-                        priority: "strict-high".into(),
-                        buffer_size_bytes: 64_000,
-                    },
-                ],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "expedited-forwarding".into(),
-                            scheduler: "ef-sched".into(),
-                        },
-                    ],
-                }],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-            }),
-            ..Default::default()
-        };
 
-        let forwarding = build_forwarding_state(&snapshot);
-        let queue_id = resolve_cos_queue_id(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
 
-        assert_eq!(queue_id, Some(1));
-    }
 
-    #[test]
-    fn resolve_cached_cos_tx_selection_prefers_egress_output_filter_and_keeps_counter() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![
-                InterfaceSnapshot {
-                    name: "reth1.0".into(),
-                    ifindex: 101,
-                    parent_ifindex: 5,
-                    vlan_id: 0,
-                    hardware_addr: "02:bf:72:00:61:01".into(),
-                    filter_input_v4: "cos-classify".into(),
-                    ..Default::default()
-                },
-                InterfaceSnapshot {
-                    name: "reth0.0".into(),
-                    ifindex: 202,
-                    hardware_addr: "02:bf:72:00:80:08".into(),
-                    filter_output_v4: "wan-classify".into(),
-                    cos_shaping_rate_bytes_per_sec: 10_000_000,
-                    cos_shaping_burst_bytes: 256_000,
-                    cos_scheduler_map: "wan-map".into(),
-                    ..Default::default()
-                },
-            ],
-            filters: vec![
-                FirewallFilterSnapshot {
-                    name: "cos-classify".into(),
-                    family: "inet".into(),
-                    terms: vec![FirewallTermSnapshot {
-                        name: "voice".into(),
-                        protocols: vec!["tcp".into()],
-                        destination_ports: vec!["443".into()],
-                        action: "accept".into(),
-                        forwarding_class: "best-effort".into(),
-                        ..Default::default()
-                    }],
-                },
-                FirewallFilterSnapshot {
-                    name: "wan-classify".into(),
-                    family: "inet".into(),
-                    terms: vec![FirewallTermSnapshot {
-                        name: "voice".into(),
-                        protocols: vec!["tcp".into()],
-                        destination_ports: vec!["443".into()],
-                        action: "accept".into(),
-                        count: "wan-hits".into(),
-                        forwarding_class: "expedited-forwarding".into(),
-                        ..Default::default()
-                    }],
-                },
-            ],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 0,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "expedited-forwarding".into(),
-                        queue: 1,
-                    },
-                ],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 4_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "ef-sched".into(),
-                        transmit_rate_bytes: 6_000_000,
-                        transmit_rate_exact: false,
-                        priority: "strict-high".into(),
-                        buffer_size_bytes: 64_000,
-                    },
-                ],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "expedited-forwarding".into(),
-                            scheduler: "ef-sched".into(),
-                        },
-                    ],
-                }],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-            }),
-            ..Default::default()
-        };
 
-        let forwarding = build_forwarding_state(&snapshot);
-        let cached = resolve_cached_cos_tx_selection(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
 
-        assert_eq!(cached.queue_id, Some(1));
-        assert_eq!(cached.dscp_rewrite, None);
-        assert!(cached.filter_counter.is_some());
-    }
 
-    #[test]
-    fn resolve_cos_queue_id_uses_ingress_input_filter_when_no_output_filter_exists() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![
-                InterfaceSnapshot {
-                    name: "reth1.0".into(),
-                    ifindex: 101,
-                    parent_ifindex: 5,
-                    vlan_id: 0,
-                    hardware_addr: "02:bf:72:00:61:01".into(),
-                    filter_input_v4: "cos-classify".into(),
-                    ..Default::default()
-                },
-                InterfaceSnapshot {
-                    name: "reth0.0".into(),
-                    ifindex: 202,
-                    hardware_addr: "02:bf:72:00:80:08".into(),
-                    cos_shaping_rate_bytes_per_sec: 10_000_000,
-                    cos_shaping_burst_bytes: 256_000,
-                    cos_scheduler_map: "wan-map".into(),
-                    ..Default::default()
-                },
-            ],
-            filters: vec![FirewallFilterSnapshot {
-                name: "cos-classify".into(),
-                family: "inet".into(),
-                terms: vec![FirewallTermSnapshot {
-                    name: "voice".into(),
-                    protocols: vec!["tcp".into()],
-                    destination_ports: vec!["443".into()],
-                    action: "accept".into(),
-                    forwarding_class: "expedited-forwarding".into(),
-                    ..Default::default()
-                }],
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 0,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "expedited-forwarding".into(),
-                        queue: 1,
-                    },
-                ],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 4_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "ef-sched".into(),
-                        transmit_rate_bytes: 6_000_000,
-                        transmit_rate_exact: false,
-                        priority: "strict-high".into(),
-                        buffer_size_bytes: 64_000,
-                    },
-                ],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "expedited-forwarding".into(),
-                            scheduler: "ef-sched".into(),
-                        },
-                    ],
-                }],
-            }),
-            ..Default::default()
-        };
 
-        let forwarding = build_forwarding_state(&snapshot);
-        let queue_id = resolve_cos_queue_id(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
 
-        assert_eq!(queue_id, Some(1));
-    }
 
-    #[test]
-    fn resolve_cached_cos_tx_selection_uses_ingress_input_filter_when_no_output_exists() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![
-                InterfaceSnapshot {
-                    name: "reth1.0".into(),
-                    ifindex: 101,
-                    parent_ifindex: 5,
-                    vlan_id: 0,
-                    hardware_addr: "02:bf:72:00:61:01".into(),
-                    filter_input_v4: "cos-classify".into(),
-                    ..Default::default()
-                },
-                InterfaceSnapshot {
-                    name: "reth0.0".into(),
-                    ifindex: 202,
-                    hardware_addr: "02:bf:72:00:80:08".into(),
-                    cos_shaping_rate_bytes_per_sec: 10_000_000,
-                    cos_shaping_burst_bytes: 256_000,
-                    cos_scheduler_map: "wan-map".into(),
-                    ..Default::default()
-                },
-            ],
-            filters: vec![FirewallFilterSnapshot {
-                name: "cos-classify".into(),
-                family: "inet".into(),
-                terms: vec![FirewallTermSnapshot {
-                    name: "voice".into(),
-                    protocols: vec!["tcp".into()],
-                    destination_ports: vec!["443".into()],
-                    action: "accept".into(),
-                    count: "lan-hits".into(),
-                    forwarding_class: "expedited-forwarding".into(),
-                    ..Default::default()
-                }],
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 0,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "expedited-forwarding".into(),
-                        queue: 1,
-                    },
-                ],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 4_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "ef-sched".into(),
-                        transmit_rate_bytes: 6_000_000,
-                        transmit_rate_exact: false,
-                        priority: "strict-high".into(),
-                        buffer_size_bytes: 64_000,
-                    },
-                ],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "expedited-forwarding".into(),
-                            scheduler: "ef-sched".into(),
-                        },
-                    ],
-                }],
-            }),
-            ..Default::default()
-        };
 
-        let forwarding = build_forwarding_state(&snapshot);
-        let cached = resolve_cached_cos_tx_selection(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
 
-        assert_eq!(cached.queue_id, Some(1));
-        assert_eq!(cached.dscp_rewrite, None);
-        assert!(cached.filter_counter.is_some());
-    }
 
-    #[test]
-    fn resolve_cached_cos_tx_selection_keeps_counter_only_output_filter_hits() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![InterfaceSnapshot {
-                name: "reth0.0".into(),
-                ifindex: 202,
-                hardware_addr: "02:bf:72:00:80:08".into(),
-                filter_output_v4: "wan-count".into(),
-                cos_shaping_rate_bytes_per_sec: 10_000_000,
-                cos_shaping_burst_bytes: 256_000,
-                cos_scheduler_map: "wan-map".into(),
-                ..Default::default()
-            }],
-            filters: vec![FirewallFilterSnapshot {
-                name: "wan-count".into(),
-                family: "inet".into(),
-                terms: vec![FirewallTermSnapshot {
-                    name: "count-only".into(),
-                    protocols: vec!["tcp".into()],
-                    destination_ports: vec!["443".into()],
-                    action: "accept".into(),
-                    count: "wan-hits".into(),
-                    ..Default::default()
-                }],
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![CoSForwardingClassSnapshot {
-                    name: "best-effort".into(),
-                    queue: 0,
-                }],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![CoSSchedulerSnapshot {
-                    name: "be-sched".into(),
-                    transmit_rate_bytes: 4_000_000,
-                    transmit_rate_exact: false,
-                    priority: "low".into(),
-                    buffer_size_bytes: 128_000,
-                }],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![CoSSchedulerMapEntrySnapshot {
-                        forwarding_class: "best-effort".into(),
-                        scheduler: "be-sched".into(),
-                    }],
-                }],
-            }),
-            ..Default::default()
-        };
 
-        let forwarding = build_forwarding_state(&snapshot);
-        let cached = resolve_cached_cos_tx_selection(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
 
-        assert_eq!(cached.queue_id, Some(0));
-        assert_eq!(cached.dscp_rewrite, None);
-        assert!(cached.filter_counter.is_some());
-    }
-
-    #[test]
-    fn resolve_cos_tx_selection_counts_counter_only_output_filter_hits() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![InterfaceSnapshot {
-                name: "reth0.0".into(),
-                ifindex: 202,
-                hardware_addr: "02:bf:72:00:80:08".into(),
-                filter_output_v4: "wan-count".into(),
-                cos_shaping_rate_bytes_per_sec: 10_000_000,
-                cos_shaping_burst_bytes: 256_000,
-                cos_scheduler_map: "wan-map".into(),
-                ..Default::default()
-            }],
-            filters: vec![FirewallFilterSnapshot {
-                name: "wan-count".into(),
-                family: "inet".into(),
-                terms: vec![FirewallTermSnapshot {
-                    name: "count-only".into(),
-                    protocols: vec!["tcp".into()],
-                    destination_ports: vec!["443".into()],
-                    action: "accept".into(),
-                    count: "wan-hits".into(),
-                    ..Default::default()
-                }],
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![CoSForwardingClassSnapshot {
-                    name: "best-effort".into(),
-                    queue: 0,
-                }],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![CoSSchedulerSnapshot {
-                    name: "be-sched".into(),
-                    transmit_rate_bytes: 4_000_000,
-                    transmit_rate_exact: false,
-                    priority: "low".into(),
-                    buffer_size_bytes: 128_000,
-                }],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![CoSSchedulerMapEntrySnapshot {
-                        forwarding_class: "best-effort".into(),
-                        scheduler: "be-sched".into(),
-                    }],
-                }],
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let selection = resolve_cos_tx_selection(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                pkt_len: 1514,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        assert_eq!(selection.queue_id, Some(0));
-        assert_eq!(selection.dscp_rewrite, None);
-
-        let filter = forwarding
-            .filter_state
-            .filters
-            .get("inet:wan-count")
-            .expect("inet output filter");
-        let term = filter.terms.first().expect("first term");
-        assert_eq!(term.counter.packets.load(Ordering::Relaxed), 1);
-        assert_eq!(term.counter.bytes.load(Ordering::Relaxed), 1514);
-    }
-
-    #[test]
-    fn resolve_cos_tx_selection_uses_ingress_filter_dscp_rewrite_when_no_output_filter_exists() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![
-                InterfaceSnapshot {
-                    name: "reth1.0".into(),
-                    ifindex: 101,
-                    parent_ifindex: 5,
-                    vlan_id: 0,
-                    hardware_addr: "02:bf:72:00:61:01".into(),
-                    filter_input_v4: "cos-classify".into(),
-                    ..Default::default()
-                },
-                InterfaceSnapshot {
-                    name: "reth0.0".into(),
-                    ifindex: 202,
-                    hardware_addr: "02:bf:72:00:80:08".into(),
-                    cos_shaping_rate_bytes_per_sec: 10_000_000,
-                    cos_shaping_burst_bytes: 256_000,
-                    cos_scheduler_map: "wan-map".into(),
-                    ..Default::default()
-                },
-            ],
-            filters: vec![FirewallFilterSnapshot {
-                name: "cos-classify".into(),
-                family: "inet".into(),
-                terms: vec![FirewallTermSnapshot {
-                    name: "voice".into(),
-                    protocols: vec!["tcp".into()],
-                    destination_ports: vec!["443".into()],
-                    action: "accept".into(),
-                    forwarding_class: "expedited-forwarding".into(),
-                    dscp_rewrite: Some(0),
-                    ..Default::default()
-                }],
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 0,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "expedited-forwarding".into(),
-                        queue: 1,
-                    },
-                ],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 4_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "ef-sched".into(),
-                        transmit_rate_bytes: 6_000_000,
-                        transmit_rate_exact: false,
-                        priority: "strict-high".into(),
-                        buffer_size_bytes: 64_000,
-                    },
-                ],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "expedited-forwarding".into(),
-                            scheduler: "ef-sched".into(),
-                        },
-                    ],
-                }],
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let selection = resolve_cos_tx_selection(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 46,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        assert_eq!(selection.queue_id, Some(1));
-        assert_eq!(selection.dscp_rewrite, Some(0));
-    }
-
-    #[test]
-    fn resolve_cos_tx_selection_skips_ingress_filter_without_tx_selection_effects() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![
-                InterfaceSnapshot {
-                    name: "reth1.0".into(),
-                    ifindex: 101,
-                    parent_ifindex: 5,
-                    vlan_id: 0,
-                    hardware_addr: "02:bf:72:00:61:01".into(),
-                    filter_input_v4: "sfmix-pbr".into(),
-                    ..Default::default()
-                },
-                InterfaceSnapshot {
-                    name: "reth0.0".into(),
-                    ifindex: 202,
-                    hardware_addr: "02:bf:72:00:80:08".into(),
-                    cos_shaping_rate_bytes_per_sec: 10_000_000,
-                    cos_shaping_burst_bytes: 256_000,
-                    cos_scheduler_map: "wan-map".into(),
-                    ..Default::default()
-                },
-            ],
-            filters: vec![FirewallFilterSnapshot {
-                name: "sfmix-pbr".into(),
-                family: "inet".into(),
-                terms: vec![FirewallTermSnapshot {
-                    name: "sfmix-route".into(),
-                    protocols: vec!["tcp".into()],
-                    destination_ports: vec!["443".into()],
-                    action: "accept".into(),
-                    count: "tx-duplicate".into(),
-                    routing_instance: "sfmix".into(),
-                    ..Default::default()
-                }],
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![CoSForwardingClassSnapshot {
-                    name: "best-effort".into(),
-                    queue: 7,
-                }],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![CoSSchedulerSnapshot {
-                    name: "be-sched".into(),
-                    transmit_rate_bytes: 10_000_000,
-                    transmit_rate_exact: false,
-                    priority: "low".into(),
-                    buffer_size_bytes: 128_000,
-                }],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![CoSSchedulerMapEntrySnapshot {
-                        forwarding_class: "best-effort".into(),
-                        scheduler: "be-sched".into(),
-                    }],
-                }],
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let selection = resolve_cos_tx_selection(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                pkt_len: 1500,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        assert_eq!(selection.queue_id, Some(7));
-        assert_eq!(selection.dscp_rewrite, None);
-        let filter = forwarding
-            .filter_state
-            .filters
-            .get("inet:sfmix-pbr")
-            .expect("filter");
-        assert_eq!(
-            filter.terms[0]
-                .counter
-                .packets
-                .load(std::sync::atomic::Ordering::Relaxed),
-            0
-        );
-    }
-
-    #[test]
-    fn resolve_cos_tx_selection_returns_none_when_no_cos_or_tx_selection_filters_exist() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![InterfaceSnapshot {
-                name: "reth1.0".into(),
-                ifindex: 101,
-                parent_ifindex: 5,
-                vlan_id: 0,
-                hardware_addr: "02:bf:72:00:61:01".into(),
-                filter_input_v4: "sfmix-pbr".into(),
-                ..Default::default()
-            }],
-            filters: vec![FirewallFilterSnapshot {
-                name: "sfmix-pbr".into(),
-                family: "inet".into(),
-                terms: vec![FirewallTermSnapshot {
-                    name: "sfmix-route".into(),
-                    protocols: vec!["tcp".into()],
-                    destination_ports: vec!["443".into()],
-                    action: "accept".into(),
-                    count: "tx-duplicate".into(),
-                    routing_instance: "sfmix".into(),
-                    ..Default::default()
-                }],
-            }],
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let selection = resolve_cos_tx_selection(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                pkt_len: 1500,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        assert_eq!(selection.queue_id, None);
-        assert_eq!(selection.dscp_rewrite, None);
-        let filter = forwarding
-            .filter_state
-            .filters
-            .get("inet:sfmix-pbr")
-            .expect("filter");
-        assert_eq!(
-            filter.terms[0]
-                .counter
-                .packets
-                .load(std::sync::atomic::Ordering::Relaxed),
-            0
-        );
-    }
-
-    #[test]
-    fn resolve_cos_queue_id_falls_back_to_default_queue_without_filter_match() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![InterfaceSnapshot {
-                ifindex: 202,
-                hardware_addr: "02:bf:72:00:80:08".into(),
-                cos_shaping_rate_bytes_per_sec: 10_000_000,
-                cos_scheduler_map: "wan-map".into(),
-                ..Default::default()
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![CoSForwardingClassSnapshot {
-                    name: "best-effort".into(),
-                    queue: 7,
-                }],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![CoSSchedulerSnapshot {
-                    name: "be-sched".into(),
-                    transmit_rate_bytes: 10_000_000,
-                    transmit_rate_exact: false,
-                    priority: "low".into(),
-                    buffer_size_bytes: 128_000,
-                }],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![CoSSchedulerMapEntrySnapshot {
-                        forwarding_class: "best-effort".into(),
-                        scheduler: "be-sched".into(),
-                    }],
-                }],
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let queue_id = resolve_cos_queue_id(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 999,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                ..Default::default()
-            },
-            None,
-        );
-
-        assert_eq!(queue_id, Some(7));
-    }
-
-    #[test]
-    fn resolve_cos_queue_id_uses_dscp_classifier_when_filters_do_not_set_class() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![InterfaceSnapshot {
-                ifindex: 202,
-                hardware_addr: "02:bf:72:00:80:08".into(),
-                cos_shaping_rate_bytes_per_sec: 10_000_000,
-                cos_scheduler_map: "wan-map".into(),
-                cos_dscp_classifier: "wan-classifier".into(),
-                ..Default::default()
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 0,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "voice".into(),
-                        queue: 5,
-                    },
-                ],
-                dscp_classifiers: vec![CoSDSCPClassifierSnapshot {
-                    name: "wan-classifier".into(),
-                    entries: vec![CoSDSCPClassifierEntrySnapshot {
-                        forwarding_class: "voice".into(),
-                        loss_priority: "low".into(),
-                        dscp_values: vec![46],
-                    }],
-                }],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "voice".into(),
-                            scheduler: "voice-sched".into(),
-                        },
-                    ],
-                }],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 4_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "voice-sched".into(),
-                        transmit_rate_bytes: 6_000_000,
-                        transmit_rate_exact: false,
-                        priority: "strict-high".into(),
-                        buffer_size_bytes: 64_000,
-                    },
-                ],
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let queue_id = resolve_cos_queue_id(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 999,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 46,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        assert_eq!(queue_id, Some(5));
-    }
-
-    #[test]
-    fn resolve_cos_queue_id_uses_ieee8021_classifier_when_filters_do_not_set_class() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![InterfaceSnapshot {
-                ifindex: 202,
-                hardware_addr: "02:bf:72:00:80:08".into(),
-                cos_shaping_rate_bytes_per_sec: 10_000_000,
-                cos_scheduler_map: "wan-map".into(),
-                cos_ieee8021_classifier: "wan-pcp".into(),
-                ..Default::default()
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 0,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "voice".into(),
-                        queue: 5,
-                    },
-                ],
-                ieee8021_classifiers: vec![CoSIEEE8021ClassifierSnapshot {
-                    name: "wan-pcp".into(),
-                    entries: vec![CoSIEEE8021ClassifierEntrySnapshot {
-                        forwarding_class: "voice".into(),
-                        loss_priority: "low".into(),
-                        code_points: vec![5],
-                    }],
-                }],
-                dscp_rewrite_rules: vec![],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "voice".into(),
-                            scheduler: "voice-sched".into(),
-                        },
-                    ],
-                }],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 4_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "voice-sched".into(),
-                        transmit_rate_bytes: 6_000_000,
-                        transmit_rate_exact: false,
-                        priority: "strict-high".into(),
-                        buffer_size_bytes: 64_000,
-                    },
-                ],
-                ..Default::default()
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let queue_id = resolve_cos_queue_id(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 999,
-                ingress_vlan_id: 100,
-                ingress_pcp: 5,
-                ingress_vlan_present: 1,
-                addr_family: libc::AF_INET as u8,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        assert_eq!(queue_id, Some(5));
-    }
-
-    #[test]
-    fn resolve_cos_queue_id_does_not_use_ieee8021_classifier_for_untagged_packets() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![InterfaceSnapshot {
-                ifindex: 202,
-                hardware_addr: "02:bf:72:00:80:08".into(),
-                cos_shaping_rate_bytes_per_sec: 10_000_000,
-                cos_scheduler_map: "wan-map".into(),
-                cos_ieee8021_classifier: "wan-pcp".into(),
-                ..Default::default()
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 0,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "bulk".into(),
-                        queue: 3,
-                    },
-                ],
-                ieee8021_classifiers: vec![CoSIEEE8021ClassifierSnapshot {
-                    name: "wan-pcp".into(),
-                    entries: vec![CoSIEEE8021ClassifierEntrySnapshot {
-                        forwarding_class: "best-effort".into(),
-                        loss_priority: "low".into(),
-                        code_points: vec![0],
-                    }],
-                }],
-                dscp_rewrite_rules: vec![],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "bulk".into(),
-                            scheduler: "bulk-sched".into(),
-                        },
-                    ],
-                }],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 4_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "bulk-sched".into(),
-                        transmit_rate_bytes: 6_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                ],
-                ..Default::default()
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let queue_id = resolve_cos_queue_id(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 999,
-                ingress_pcp: 0,
-                ingress_vlan_present: 0,
-                addr_family: libc::AF_INET as u8,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        assert_eq!(queue_id, Some(0));
-    }
-
-    // Note on invariant change (replaces the pre-a15a6120 "defaults to iface default" behavior):
-    // The original shape of this test asserted that an output filter with NO tx-side effect (no
-    // forwarding_class, no counter) would still shadow the ingress input filter's classification
-    // and leave egress at the interface default queue.  Commit a15a6120 changed the gating so the
-    // output filter is skipped entirely when it has neither forwarding_class, dscp_rewrite, nor
-    // counter terms — matching Junos semantics, where a classify-only output filter that does not
-    // classify does not clobber upstream classification.  The new invariant asserted below: when
-    // the output filter has no tx-side effect, ingress input-filter classification is preserved.
-    #[test]
-    fn resolve_cos_queue_id_preserves_ingress_classification_when_output_filter_has_no_forwarding_class()
-     {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![
-                InterfaceSnapshot {
-                    name: "reth1.0".into(),
-                    ifindex: 101,
-                    parent_ifindex: 5,
-                    vlan_id: 0,
-                    hardware_addr: "02:bf:72:00:61:01".into(),
-                    filter_input_v4: "cos-classify".into(),
-                    ..Default::default()
-                },
-                InterfaceSnapshot {
-                    name: "reth0.0".into(),
-                    ifindex: 202,
-                    hardware_addr: "02:bf:72:00:80:08".into(),
-                    filter_output_v4: "wan-classify".into(),
-                    cos_shaping_rate_bytes_per_sec: 10_000_000,
-                    cos_shaping_burst_bytes: 256_000,
-                    cos_scheduler_map: "wan-map".into(),
-                    ..Default::default()
-                },
-            ],
-            filters: vec![
-                FirewallFilterSnapshot {
-                    name: "cos-classify".into(),
-                    family: "inet".into(),
-                    terms: vec![FirewallTermSnapshot {
-                        name: "voice".into(),
-                        protocols: vec!["tcp".into()],
-                        destination_ports: vec!["443".into()],
-                        action: "accept".into(),
-                        forwarding_class: "expedited-forwarding".into(),
-                        ..Default::default()
-                    }],
-                },
-                FirewallFilterSnapshot {
-                    name: "wan-classify".into(),
-                    family: "inet".into(),
-                    terms: vec![FirewallTermSnapshot {
-                        name: "allow".into(),
-                        protocols: vec!["tcp".into()],
-                        destination_ports: vec!["443".into()],
-                        action: "accept".into(),
-                        ..Default::default()
-                    }],
-                },
-            ],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![
-                    CoSForwardingClassSnapshot {
-                        name: "best-effort".into(),
-                        queue: 7,
-                    },
-                    CoSForwardingClassSnapshot {
-                        name: "expedited-forwarding".into(),
-                        queue: 1,
-                    },
-                ],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![
-                    CoSSchedulerSnapshot {
-                        name: "be-sched".into(),
-                        transmit_rate_bytes: 10_000_000,
-                        transmit_rate_exact: false,
-                        priority: "low".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                    CoSSchedulerSnapshot {
-                        name: "ef-sched".into(),
-                        transmit_rate_bytes: 10_000_000,
-                        transmit_rate_exact: false,
-                        priority: "strict-high".into(),
-                        buffer_size_bytes: 128_000,
-                    },
-                ],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "best-effort".into(),
-                            scheduler: "be-sched".into(),
-                        },
-                        CoSSchedulerMapEntrySnapshot {
-                            forwarding_class: "expedited-forwarding".into(),
-                            scheduler: "ef-sched".into(),
-                        },
-                    ],
-                }],
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let queue_id = resolve_cos_queue_id(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        // cos-classify on reth1.0 maps expedited-forwarding -> queue 1.  The output filter
-        // wan-classify on reth0.0 has no tx-side effect (no forwarding_class, no dscp_rewrite,
-        // no counter), so post-a15a6120 it is bypassed and the ingress classification is
-        // preserved.  Pre-a15a6120 this was expected to fall through to the iface default queue
-        // (best-effort = 7); that contract no longer holds and is captured by this test.
-        assert_eq!(queue_id, Some(1));
-    }
-
-    #[test]
-    fn resolve_cos_tx_selection_preserves_output_filter_dscp_rewrite_without_forwarding_class() {
-        let snapshot = ConfigSnapshot {
-            interfaces: vec![InterfaceSnapshot {
-                name: "reth0.0".into(),
-                ifindex: 202,
-                hardware_addr: "02:bf:72:00:80:08".into(),
-                filter_output_v4: "wan-rewrite".into(),
-                cos_shaping_rate_bytes_per_sec: 10_000_000,
-                cos_scheduler_map: "wan-map".into(),
-                ..Default::default()
-            }],
-            filters: vec![FirewallFilterSnapshot {
-                name: "wan-rewrite".into(),
-                family: "inet".into(),
-                terms: vec![FirewallTermSnapshot {
-                    name: "rewrite".into(),
-                    protocols: vec!["tcp".into()],
-                    destination_ports: vec!["443".into()],
-                    action: "accept".into(),
-                    dscp_rewrite: Some(46),
-                    ..Default::default()
-                }],
-            }],
-            class_of_service: Some(ClassOfServiceSnapshot {
-                forwarding_classes: vec![CoSForwardingClassSnapshot {
-                    name: "best-effort".into(),
-                    queue: 7,
-                }],
-                dscp_classifiers: vec![],
-                ieee8021_classifiers: vec![],
-                dscp_rewrite_rules: vec![],
-                schedulers: vec![CoSSchedulerSnapshot {
-                    name: "be-sched".into(),
-                    transmit_rate_bytes: 10_000_000,
-                    transmit_rate_exact: false,
-                    priority: "low".into(),
-                    buffer_size_bytes: 128_000,
-                }],
-                scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                    name: "wan-map".into(),
-                    entries: vec![CoSSchedulerMapEntrySnapshot {
-                        forwarding_class: "best-effort".into(),
-                        scheduler: "be-sched".into(),
-                    }],
-                }],
-            }),
-            ..Default::default()
-        };
-
-        let forwarding = build_forwarding_state(&snapshot);
-        let selection = resolve_cos_tx_selection(
-            &forwarding,
-            202,
-            UserspaceDpMeta {
-                ingress_ifindex: 5,
-                ingress_vlan_id: 0,
-                addr_family: libc::AF_INET as u8,
-                dscp: 0,
-                ..Default::default()
-            },
-            Some(&SessionKey {
-                addr_family: libc::AF_INET as u8,
-                protocol: PROTO_TCP,
-                src_ip: IpAddr::V4(Ipv4Addr::new(10, 0, 61, 100)),
-                dst_ip: IpAddr::V4(Ipv4Addr::new(172, 16, 80, 200)),
-                src_port: 12345,
-                dst_port: 443,
-            }),
-        );
-
-        assert_eq!(selection.queue_id, Some(7));
-        assert_eq!(selection.dscp_rewrite, Some(46));
-    }
 
     #[test]
     fn assign_local_dscp_rewrite_preserves_existing_filter_rewrite() {
@@ -7144,344 +5272,6 @@ mod tests {
         assert_eq!(
             after.tx_ring_full_submit_stalls,
             before.tx_ring_full_submit_stalls
-        );
-    }
-
-
-    // ---------------------------------------------------------------------
-    // #718 ECN CE-marking. The markers are the load-bearing helpers;
-    // the admission-path tests exercise `apply_cos_admission_ecn_policy`
-    // which is what `enqueue_cos_item` calls in-line. Keep the marker
-    // tests byte-precise so a future refactor that flips an endian /
-    // offset / masks a different bit fails loudly.
-    // ---------------------------------------------------------------------
-
-
-
-
-
-    #[test]
-    fn mark_ecn_ce_ipv4_converts_ect0_to_ce_and_updates_checksum() {
-        // ECT(0) = 0b10 in the low 2 bits of the TOS byte. Pick a
-        // non-zero DSCP (0x28 = CS5 = expedited forwarding) to verify
-        // the upper 6 bits survive the mark. TOS before = 0xa2.
-        let tos = (0x28u8 << 2) | ECN_ECT_0;
-        let mut pkt = build_ipv4_test_packet(tos);
-        assert_eq!(ipv4_tos(&pkt), 0xa2);
-        let csum_before = ipv4_checksum(&pkt);
-
-        assert!(mark_ecn_ce_ipv4(&mut pkt, 14));
-
-        // Low 2 bits now CE, upper 6 bits (DSCP) unchanged.
-        assert_eq!(ipv4_tos(&pkt) & ECN_MASK, ECN_CE);
-        assert_eq!(ipv4_tos(&pkt) >> 2, 0x28);
-        // Checksum must differ from the before-state (ECN flipped one
-        // bit in the low byte) AND be valid from scratch.
-        assert_ne!(
-            ipv4_checksum(&pkt),
-            csum_before,
-            "ECN bit flip must change the IP checksum",
-        );
-        assert_eq!(
-            ipv4_checksum(&pkt),
-            compute_ipv4_header_checksum(&pkt[14..34]),
-            "incremental checksum must match a from-scratch recompute",
-        );
-    }
-
-    #[test]
-    fn mark_ecn_ce_ipv4_converts_ect1_to_ce_and_updates_checksum() {
-        // ECT(1) = 0b01. DSCP = 0, so TOS starts at 0x01 — stresses
-        // the case where the high nibble is zero and only the low
-        // bits mutate.
-        let tos = ECN_ECT_1;
-        let mut pkt = build_ipv4_test_packet(tos);
-
-        assert!(mark_ecn_ce_ipv4(&mut pkt, 14));
-        assert_eq!(ipv4_tos(&pkt), ECN_CE);
-        assert_eq!(
-            ipv4_checksum(&pkt),
-            compute_ipv4_header_checksum(&pkt[14..34]),
-        );
-    }
-
-    #[test]
-    fn mark_ecn_ce_ipv4_leaves_not_ect_untouched() {
-        // NOT-ECT packet must be left entirely alone — RFC 3168 6.1.1.1
-        // forbids forcing ECN on flows that did not negotiate it.
-        let tos = 0xb8; // DSCP 46 (EF), ECN = 00
-        let mut pkt = build_ipv4_test_packet(tos);
-        let before = pkt.clone();
-
-        assert!(!mark_ecn_ce_ipv4(&mut pkt, 14));
-        assert_eq!(pkt, before, "NOT-ECT packet must be byte-identical");
-    }
-
-    #[test]
-    fn mark_ecn_ce_ipv4_leaves_ce_untouched() {
-        // CE already — idempotent: function reports "not marked" but
-        // also doesn't re-write the checksum, so bytes stay identical.
-        let tos = 0xb8 | ECN_CE;
-        let mut pkt = build_ipv4_test_packet(tos);
-        let before = pkt.clone();
-
-        assert!(!mark_ecn_ce_ipv4(&mut pkt, 14));
-        assert_eq!(pkt, before, "CE packet must be byte-identical");
-    }
-
-    #[test]
-    fn mark_ecn_ce_ipv4_rejects_short_buffer() {
-        // Buffer too short to hold a full 20-byte IPv4 header starting
-        // at l3_offset=14 (only 33 bytes — one short). Must return
-        // false and not panic.
-        let mut pkt = vec![0u8; 33];
-        assert!(!mark_ecn_ce_ipv4(&mut pkt, 14));
-
-        // Also exercise the case where `l3_offset` itself pushes past
-        // the buffer end.
-        let mut pkt = vec![0u8; 16];
-        assert!(!mark_ecn_ce_ipv4(&mut pkt, 14));
-    }
-
-
-
-    #[test]
-    fn mark_ecn_ce_ipv6_converts_ect0_to_ce() {
-        // DSCP 46 (EF) + ECT(0) → full tclass 0xba.
-        let tclass = (0x2eu8 << 2) | ECN_ECT_0;
-        let mut pkt = build_ipv6_test_packet(tclass);
-        assert_eq!(ipv6_tclass(&pkt), 0xba);
-        // Preserve flow label / version bits for the round-trip check.
-        let version_nibble_before = pkt[14] & 0xf0;
-        let flow_label_low_before = pkt[15] & 0x0f;
-
-        assert!(mark_ecn_ce_ipv6(&mut pkt, 14));
-        assert_eq!(ipv6_tclass(&pkt) & ECN_MASK, ECN_CE);
-        assert_eq!(ipv6_tclass(&pkt) >> 2, 0x2e);
-        // Version + flow-label bits must not drift.
-        assert_eq!(pkt[14] & 0xf0, version_nibble_before);
-        assert_eq!(pkt[15] & 0x0f, flow_label_low_before);
-    }
-
-    #[test]
-    fn mark_ecn_ce_ipv6_converts_ect1_to_ce() {
-        let tclass = ECN_ECT_1;
-        let mut pkt = build_ipv6_test_packet(tclass);
-        assert!(mark_ecn_ce_ipv6(&mut pkt, 14));
-        assert_eq!(ipv6_tclass(&pkt), ECN_CE);
-    }
-
-    #[test]
-    fn mark_ecn_ce_ipv6_leaves_not_ect_untouched() {
-        let tclass = 0xb8; // DSCP 46, ECN 00
-        let mut pkt = build_ipv6_test_packet(tclass);
-        let before = pkt.clone();
-        assert!(!mark_ecn_ce_ipv6(&mut pkt, 14));
-        assert_eq!(pkt, before);
-    }
-
-    #[test]
-    fn mark_ecn_ce_ipv6_leaves_ce_untouched() {
-        let tclass = 0xb8 | ECN_CE;
-        let mut pkt = build_ipv6_test_packet(tclass);
-        let before = pkt.clone();
-        assert!(!mark_ecn_ce_ipv6(&mut pkt, 14));
-        assert_eq!(pkt, before);
-    }
-
-    #[test]
-    fn mark_ecn_ce_ipv6_rejects_short_buffer() {
-        let mut pkt = vec![0u8; 15];
-        assert!(!mark_ecn_ce_ipv6(&mut pkt, 14));
-    }
-
-    #[test]
-    fn maybe_mark_ecn_ce_dispatches_by_addr_family() {
-        // IPv4 dispatch: ECT(0) → CE.
-        let tos = ECN_ECT_0;
-        let bytes = build_ipv4_test_packet(tos);
-        let mut req = TxRequest {
-            bytes,
-            expected_ports: None,
-            expected_addr_family: libc::AF_INET as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: None,
-            egress_ifindex: 1,
-            cos_queue_id: Some(0),
-            dscp_rewrite: None,
-        };
-        assert!(maybe_mark_ecn_ce(&mut req));
-        assert_eq!(req.bytes[15] & ECN_MASK, ECN_CE);
-
-        // IPv6 dispatch: ECT(1) → CE.
-        let tclass = ECN_ECT_1;
-        let bytes = build_ipv6_test_packet(tclass);
-        let mut req = TxRequest {
-            bytes,
-            expected_ports: None,
-            expected_addr_family: libc::AF_INET6 as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: None,
-            egress_ifindex: 1,
-            cos_queue_id: Some(0),
-            dscp_rewrite: None,
-        };
-        assert!(maybe_mark_ecn_ce(&mut req));
-        assert_eq!(ipv6_tclass(&req.bytes), ECN_CE);
-
-        // Unknown address family: no-op (and no panic).
-        let mut req = TxRequest {
-            bytes: vec![0u8; 64],
-            expected_ports: None,
-            expected_addr_family: 0,
-            expected_protocol: PROTO_TCP,
-            flow_key: None,
-            egress_ifindex: 1,
-            cos_queue_id: Some(0),
-            dscp_rewrite: None,
-        };
-        assert!(!maybe_mark_ecn_ce(&mut req));
-    }
-
-    /// Regression pin for the VLAN-tagged admission path discovered in
-    /// the #727 live validation: a single 802.1Q tag (ethertype 0x8100)
-    /// pushes L3 four bytes deeper. `maybe_mark_ecn_ce` must detect
-    /// that via `ethernet_l3_offset` and still mark the ECN bits at
-    /// the correct offset rather than stamping into the VLAN TCI.
-    #[test]
-    fn maybe_mark_ecn_ce_handles_single_vlan_tagged_frame() {
-        // Build a standard IPv4 test packet, then splice a 4-byte VLAN
-        // tag between the MAC addresses and the ethertype. The result
-        // is: 6 dst + 6 src + TPID(0x8100) + TCI(VID=80, prio=5) +
-        //     EthType(0x0800) + <20-byte IPv4 header>.
-        let tos = ECN_ECT_0;
-        let base = build_ipv4_test_packet(tos);
-        let mut tagged = Vec::with_capacity(base.len() + 4);
-        tagged.extend_from_slice(&base[..12]); // dst + src MAC
-        tagged.extend_from_slice(&[0x81, 0x00]); // TPID
-        // TCI: priority 5 << 13 | DEI 0 | VID 80.
-        let tci: u16 = (5 << 13) | 80;
-        tagged.extend_from_slice(&tci.to_be_bytes());
-        tagged.extend_from_slice(&[0x08, 0x00]); // inner ethertype (IPv4)
-        tagged.extend_from_slice(&base[14..]); // IPv4 header + payload
-
-        // Confirm `ethernet_l3` parses IPv4 at offset 18 for this frame.
-        assert_eq!(ethernet_l3(&tagged), Some(EthernetL3::Ipv4(18)));
-
-        let mut req = TxRequest {
-            bytes: tagged,
-            expected_ports: None,
-            expected_addr_family: libc::AF_INET as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: None,
-            egress_ifindex: 1,
-            cos_queue_id: Some(4),
-            dscp_rewrite: None,
-        };
-        assert!(
-            maybe_mark_ecn_ce(&mut req),
-            "VLAN-tagged ECT(0) frame must be marked at the VLAN-shifted L3 offset"
-        );
-        // TOS byte sits at l3_offset + 1 = 19 in the tagged frame.
-        assert_eq!(req.bytes[19] & ECN_MASK, ECN_CE);
-        // And critically: the VLAN TCI bytes must NOT have been
-        // mutated — if the old hardcoded offset 14 had hit, the "ECN
-        // bits" we'd have touched are inside the VLAN priority nibble
-        // at byte 15, which we assert stayed intact.
-        let tci_after = u16::from_be_bytes([req.bytes[14], req.bytes[15]]);
-        assert_eq!(
-            tci_after, tci,
-            "VLAN TCI must be untouched by ECN marking"
-        );
-    }
-
-    /// Counter-factual: ethertype 0 (or anything we don't understand)
-    /// returns `None` from `ethernet_l3`, so marking is a no-op.
-    /// Guards against a regression that defaults to offset 14 on
-    /// unknown frames.
-    #[test]
-    fn maybe_mark_ecn_ce_rejects_unknown_ethertype() {
-        let mut req = TxRequest {
-            bytes: {
-                let mut b = build_ipv4_test_packet(ECN_ECT_0);
-                b[12] = 0x12;
-                b[13] = 0x34;
-                b
-            },
-            expected_ports: None,
-            expected_addr_family: libc::AF_INET as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: None,
-            egress_ifindex: 1,
-            cos_queue_id: Some(0),
-            dscp_rewrite: None,
-        };
-        assert_eq!(ethernet_l3(&req.bytes), None);
-        assert!(!maybe_mark_ecn_ce(&mut req));
-        // ECT(0) bits at the would-have-been-wrong-offset untouched.
-        assert_eq!(req.bytes[15] & ECN_MASK, ECN_ECT_0);
-    }
-
-    /// QinQ (0x88A8 outer + 0x8100 inner) must be rejected rather than
-    /// guessed at, because L3 actually lives at offset 22 on those
-    /// frames and a default to 18 would stamp into the inner VLAN TCI.
-    /// #728 review pin: once we've paid to parse the outer ethertype,
-    /// the parse must be the source of truth.
-    #[test]
-    fn ethernet_l3_rejects_qinq_until_explicitly_supported() {
-        let base = build_ipv4_test_packet(ECN_ECT_0);
-        let mut qinq = Vec::with_capacity(base.len() + 8);
-        qinq.extend_from_slice(&base[..12]); // MACs
-        // Outer 802.1ad: TPID 0x88A8, TCI with an outer VID 100.
-        qinq.extend_from_slice(&[0x88, 0xA8]);
-        let outer_tci: u16 = 100;
-        qinq.extend_from_slice(&outer_tci.to_be_bytes());
-        // Inner 802.1Q: TPID 0x8100 at the "inner ethertype" position.
-        qinq.extend_from_slice(&[0x81, 0x00]);
-        let inner_tci: u16 = 80;
-        qinq.extend_from_slice(&inner_tci.to_be_bytes());
-        qinq.extend_from_slice(&[0x08, 0x00]); // IPv4 (well beyond where we care)
-        qinq.extend_from_slice(&base[14..]);
-
-        assert_eq!(
-            ethernet_l3(&qinq),
-            None,
-            "QinQ (0x88A8 → 0x8100) must be rejected — inner VLAN tag not yet supported"
-        );
-
-        // And the marker refuses such a frame — no ECN bits are flipped.
-        let mut req = TxRequest {
-            bytes: qinq,
-            expected_ports: None,
-            expected_addr_family: libc::AF_INET as u8,
-            expected_protocol: PROTO_TCP,
-            flow_key: None,
-            egress_ifindex: 1,
-            cos_queue_id: Some(4),
-            dscp_rewrite: None,
-        };
-        assert!(!maybe_mark_ecn_ce(&mut req));
-    }
-
-    /// A VLAN-tagged frame whose inner ethertype is ARP / MPLS / etc.
-    /// must be rejected too, matching the `refuse to guess` contract.
-    /// Without this check we'd treat offset 18 as an IPv4 TOS byte and
-    /// stamp the low 2 bits of whatever is there (ARP's hardware type
-    /// in this case), corrupting the frame.
-    #[test]
-    fn ethernet_l3_rejects_vlan_tagged_non_ip_payload() {
-        let base = build_ipv4_test_packet(ECN_ECT_0);
-        let mut tagged = Vec::with_capacity(base.len() + 4);
-        tagged.extend_from_slice(&base[..12]);
-        tagged.extend_from_slice(&[0x81, 0x00]); // outer 802.1Q
-        let tci: u16 = 80;
-        tagged.extend_from_slice(&tci.to_be_bytes());
-        tagged.extend_from_slice(&[0x08, 0x06]); // inner = ARP (0x0806)
-        tagged.extend_from_slice(&base[14..]);
-        assert_eq!(
-            ethernet_l3(&tagged),
-            None,
-            "VLAN-tagged non-IP payload must not dispatch to an IP marker",
         );
     }
